define configuration functions as inline members in opt_parser
